newupdate

+++ src/components/singleBolg.vue
@@ -1,6 +1,11 @@
 <template lang="html">
   <div id="single-blog" class="mt-3">
-    <b-button variant="danger" class="mb-5" v-on:click="destroyed">Destroy blog</b-button>
+    <div class="">
+      <b-button variant="danger" class="mb-5 " v-on:click="destroyed">Destroy blog</b-button>
+      <b-button to='/add'  class="mb-5 ml-2" variant="success" exact>Edit blog</b-button>
+
+    </div>
+
     <h1>{{blog.title}}</h1>
     <article >
       {{blog.content}}
@@ -31,7 +36,7 @@ export default {
  },
  methods:{
    destroyed: function(){
-     alert('are you sure')
+     if(confirm('are you sure?'))
      this.$http.delete('https://myrailblog-default-rtdb.europe-west1.firebasedatabase.app/post/' + this.id + '.json').then(function(){
        return this.$router.push({path: '/'});
      })
@@ -45,7 +50,7 @@ export default {
     max-width: 960px;
     margin: 0 auto;
     padding: 20px;
-    background: #eee;
     border: 1px dotted #aaa;
+    background: #F5F5F5;
 }
 </style>
